OK-513: Järjestelty tyypityksiä
