+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulos/sijoittelu-status-change-button.test.tsx
@@ -2,9 +2,9 @@ import { describe, expect, test, vi } from 'vitest';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { SijoitteluStatusChangeButton } from './sijoittelu-status-change-button';
 import { useSijoitteluStatusMutation } from './useSijoitteluStatusMutation';
-import { LaskettuJonoWithHakijaInfot } from '@/app/lib/valintalaskenta-service';
 import { UserPermissions } from '@/app/lib/permissions';
 import { OPH_ORGANIZATION_OID } from '@/app/lib/constants';
+import { LaskettuJonoWithHakijaInfo } from '@/app/hooks/useLasketutValinnanVaiheet';
 
 vi.mock('@/app/hooks/useTranslations', () => ({
   useTranslations: () => {
@@ -40,7 +40,7 @@ const renderSijoitteluButton = ({
           valmisSijoiteltavaksi: true,
           siirretaanSijoitteluun: true,
           ...jono,
-        } as LaskettuJonoWithHakijaInfot
+        } as LaskettuJonoWithHakijaInfo
       }
       permissions={
         {
