+++ src/app/hooks/useLasketutValinnanVaiheet.ts
@@ -1,7 +1,29 @@
 'use client';
 import { useSuspenseQueries } from '@tanstack/react-query';
-import { getLasketutValinnanVaiheet } from '../lib/valintalaskenta-service';
+import {
+  JonoSija,
+  LaskettuValintatapajono,
+  getLasketutValinnanVaiheet,
+} from '../lib/valintalaskenta-service';
 import { Hakemus, getHakemukset } from '../lib/ataru';
+import { TranslatedName } from '../lib/localization/localization-types';
+
+export type JonoSijaWithHakijaInfo = Omit<
+  JonoSija,
+  'jarjestyskriteerit' | 'harkinnanvarainen'
+> & {
+  hakijanNimi: string;
+  hakemusOid: string;
+  henkiloOid: string;
+  pisteet?: number;
+  hakutoiveNumero?: number;
+  tuloksenTila?: string;
+  muutoksenSyy?: TranslatedName;
+};
+
+export type LaskettuJonoWithHakijaInfo = LaskettuValintatapajono & {
+  jonosijat: Array<JonoSijaWithHakijaInfo>;
+};
 
 export const useLasketutValinnanVaiheet = ({
   hakuOid,
